+++ lib/dart_eticon_api.dart
@@ -1,1 +1,1 @@
-export 'src/dart_eticon_api.dart';
+export 'src/old_api/dart_eticon_api.dart';