+++ lib/eticon_api.dart
@@ -2,5 +2,11 @@
 
 library eticon_api;
 
-export 'src/flutter_eticon_api.dart';
+export 'src/old_api/flutter_eticon_api.dart';
 export 'src/api_errors.dart';
+export 'src/flutter_eticon_api.dart';
+export 'package:dio/src/multipart_file.dart';
+export 'package:dio/src/form_data.dart';
+export 'package:dio/src/headers.dart';
+export 'package:dio/src/cancel_token.dart';
+export 'package:dio/src/response.dart';