+++ lib/src/flutter_eticon_api.dart
@@ -1,9 +1,9 @@
+import 'package:eticon_api/eticon_api.dart';
+import 'package:eticon_api/src/api_errors.dart';
+import 'package:eticon_api/src/dio_api_st.dart';
+import 'package:eticon_api/src/old_api/api_st.dart';
+import 'package:eticon_api/src/token.dart';
 import 'package:get_storage/get_storage.dart';
-
-import 'api_errors.dart';
-import 'api_st.dart';
-import 'token.dart';
-import 'type.dart';
 
 class Api {
   ///Initialization API class
@@ -15,16 +15,18 @@ class Api {
       bool enableUtf8Decoding = false,
       bool loadTokenFromMemory = true,
       String? authTitle,
+      bool usePrintInLogs = false,
+      Function(APIException error)? onAllError,
       String? storageUrl}) async {
-    if (!ApiST.instance.setInitState()) {
-      throw EticonApiError(error: 'API class already initialization');
+    if (!DioApiST.instance.setInitState()) {
+      throw APIException(-1, body: 'API class already initialization');
     }
     if (urls.isEmpty) {
-      throw EticonApiError(error: 'URLs list is empty');
+      throw APIException(-1, body: 'URLs list is empty');
     }
-    if (!urls[0].startsWith('http')) throw EticonApiError(error: 'The url should start with https or http');
+    if (!urls[0].startsWith('http')) throw APIException(-1, body: 'The url should start with https or http');
     for (int i = 0; i < urls.length; i++) {
-      if (!urls[i].endsWith('/')){
+      if (!urls[i].endsWith('/')) {
         urls[i] += '/';
       }
     }
@@ -49,55 +51,68 @@ class Api {
       Token.instance.setToken('');
       Token.instance.setRefreshToken('');
     }
+
+    ///Init new Version
+    if (authTitle != null) {
+      DioApiST.instance.setAuthTitle(authTitle);
+    }
+    DioApiST.instance.usePrint = usePrintInLogs;
+    DioApiST.instance.onAllError = onAllError;
+    DioApiST.instance.setBaseUrl(urls);
+    DioApiST.instance.setGlobalTestMode(globalTestMode);
+    DioApiST.instance.disableAllTestMode(disableAllTestMode);
+    DioApiST.instance.enableUtf8Decoding(enableUtf8Decoding);
+    DioApiST.instance.setBearerMode(bearerToken);
+    if (storageUrl != null) {
+      DioApiST.instance.setStorageUrl(storageUrl);
+    }
+
+    ///Init OldVersion
     if (authTitle != null) {
       ApiST.instance.setAuthTitle(authTitle);
     }
     ApiST.instance.setUrls(urls);
     ApiST.instance.setGlobalTestMode(globalTestMode);
     ApiST.instance.disableAllTestMode(disableAllTestMode);
-    ApiST.instance.enableUtf8Decoding(enableUtf8Decoding);
+    // ApiST.instance.enableUtf8Decoding(enableUtf8Decoding);
     ApiST.instance.setBearerMode(bearerToken);
     if (storageUrl != null) {
       ApiST.instance.setStorageUrl(storageUrl);
     }
   }
 
-  // static get baseUrl => ApiST.instance.baseUrl;
+  static get urls => DioApiST.instance.urls;
 
   ///Help to get url to resource in server storage
   static String dataFromStorage(String path) {
-    if (ApiST.instance.storageUrl == null) {
-      throw EticonApiError(error: 'Storage url is null, set storageUrl in Api.init()');
+    if (DioApiST.instance.storageUrl == null) {
+      throw APIException(-1, body: 'Storage url is null, set storageUrl in Api.init()');
     } else {
-      return '${ApiST.instance.storageUrl}$path';
+      return '${DioApiST.instance.storageUrl}$path';
     }
   }
 
   ///Set headers. Default is only "Content-type": application/json
   static void setHeaders(Map<String, String> headers) {
-    ApiST.instance.setHeaders(headers);
+    DioApiST.instance.setHeaders(headers);
   }
 
   ///Return Authorization token
   static String? get token => Token.instance.token;
 
   ///Checks token storage for emptiness
-  static bool tokenIsNotEmpty() {
-    return Token.instance.token.isNotEmpty;
-  }
+  static bool get tokenIsNotEmpty => Token.instance.token.isNotEmpty;
 
   ///Checks token storage for emptiness
-  static bool tokenIsEmpty() {
-    return Token.instance.token.isEmpty;
-  }
+  static bool get tokenIsEmpty => Token.instance.token.isEmpty;
 
   ///Clear Token
-  static void clearToken() => Token.instance.setToken('');
+  static void get clearToken => Token.instance.setToken('');
 
   ///Set Authorization token
   static void setToken(String token) {
-    if (!ApiST.instance.initState) {
-      throw EticonApiError(error: 'Need Api.init() before setToken');
+    if (!DioApiST.instance.initState) {
+      throw APIException(-1, body: 'Need Api.init() before setToken');
     }
     GetStorage().write('ApiEticonMainAuthToken2312', token);
     Token.instance.setToken(token);
@@ -110,22 +125,18 @@ class Api {
   static DateTime? get expireDate => Token.instance.expireDate.year == 1970 ? null : Token.instance.expireDate;
 
   ///Clear Token
-  static void clearRefreshToken() => Token.instance.setRefreshToken('');
+  static void get clearRefreshToken => Token.instance.setRefreshToken('');
 
   ///Checks token storage for emptiness
-  static bool refreshTokenIsNotEmpty() {
-    return Token.instance.token.isNotEmpty;
-  }
+  static bool get refreshTokenIsNotEmpty => Token.instance.token.isNotEmpty;
 
   ///Checks token storage for emptiness
-  static bool refreshTokenIsEmpty() {
-    return Token.instance.token.isEmpty;
-  }
+  static bool get refreshTokenIsEmpty => Token.instance.token.isEmpty;
 
   ///Set refresh token
   static void setRefreshToken(String token) {
-    if (!ApiST.instance.initState) {
-      throw EticonApiError(error: 'Need Api.init() before setRefreshToken');
+    if (!DioApiST.instance.initState) {
+      throw APIException(-1, body: 'Need Api.init() before setRefreshToken');
     }
     GetStorage().write('ApiEticonMainRefreshToken2312', token);
     Token.instance.setRefreshToken(token);
@@ -133,8 +144,8 @@ class Api {
 
   ///Set refresh token expire in seconds. Fro check you can use Api.isRefreshTokenExpire
   static void setExpire({required int seconds}) {
-    if (!ApiST.instance.initState) {
-      throw EticonApiError(error: 'Need Api.init() before setRefreshToken');
+    if (!DioApiST.instance.initState) {
+      throw APIException(-1, body: 'Need Api.init() before setRefreshToken');
     }
     Token.instance.setExpire(seconds);
     GetStorage().write('ApiEticonMainExpireDate2312', Token.instance.expireDate.toString());
@@ -144,144 +155,169 @@ class Api {
   static bool get isTokenExpire => DateTime.now().isAfter(Token.instance.expireDate);
 
   /// Sends an HTTP GET request.
-  static Future<Map<String, dynamic>> get(
-      {required String method,
-      bool isAuth = false,
+  static Future<dynamic> get(String path,
+      {bool isAuth = false,
       bool testMode = false,
       Map<String, dynamic>? query,
+      Map<String, String>? headers,
+      CancelToken? cancelToken,
+      ResponseType responseType = ResponseType.map_data,
+      Function(int, int)? onReceiveProgress,
+      bool ignoreOnAllError = false,
       int urlIndex = 0}) async {
-    if (ApiST.instance.urls.isEmpty) {
-      throw EticonApiError(error: 'Base url not set, use Api.init()');
+    String error = _checkBeforeRequest(isAuth, urlIndex);
+    if (error.isNotEmpty) {
+      throw APIException(-1, body: error);
     }
-    if (isAuth) {
-      if (Token.instance.token.isEmpty) {
-        throw EticonApiError(error: 'Authentication token is empty, use Api.setToken (String url)');
-      }
-    }
-    return await ApiST.instance
-        .request(type: TYPE.GET, method: method, isAuth: isAuth, testMode: testMode, query: query, urlIndex: urlIndex);
+    bool isUrl = path.startsWith('http');
+    return await DioApiST.instance.request(
+        type: 'GET',
+        method: !isUrl ? path : null,
+        baseUrl: isUrl ? path : null,
+        isAuth: isAuth,
+        testMode: testMode,
+        headers: headers,
+        responseType: responseType,
+        cancelToken: cancelToken,
+        data: query,
+        onReceiveProgress: onReceiveProgress,
+        ignoreOnAllError: ignoreOnAllError,
+        urlIndex: urlIndex);
   }
 
   /// Sends an HTTP POST request.
-  static Future<Map<String, dynamic>> post(
-      {required String method,
-      bool isAuth = false,
+  static Future<dynamic> post(String path,
+      {bool isAuth = false,
       bool testMode = false,
-      required Map<String, dynamic> body,
+      dynamic body,
+      Map<String, String>? headers,
+      CancelToken? cancelToken,
+      ResponseType responseType = ResponseType.map_data,
+      Function(int, int)? onSendProgress,
+      Function(int, int)? onReceiveProgress,
+      bool ignoreOnAllError = false,
       int urlIndex = 0}) async {
     String error = _checkBeforeRequest(isAuth, urlIndex);
     if (error.isNotEmpty) {
-      throw EticonApiError(error: error);
+      throw APIException(-1, body: error);
     }
-    return await ApiST.instance
-        .request(type: TYPE.POST, method: method, isAuth: isAuth, testMode: testMode, query: body, urlIndex: urlIndex);
+    bool isUrl = path.startsWith('http');
+    return await DioApiST.instance.request(
+        type: 'POST',
+        method: !isUrl ? path : null,
+        baseUrl: isUrl ? path : null,
+        isAuth: isAuth,
+        testMode: testMode,
+        headers: headers,
+        responseType: responseType,
+        cancelToken: cancelToken,
+        data: body,
+        onReceiveProgress: onReceiveProgress,
+        onSendProgress: onSendProgress,
+        ignoreOnAllError: ignoreOnAllError,
+        urlIndex: urlIndex);
   }
 
   /// Sends an HTTP PUT request.
-  static Future<Map<String, dynamic>> put(
-      {required String method,
-      bool isAuth = false,
+  static Future<dynamic> put(String path,
+      {bool isAuth = false,
       bool testMode = false,
-      required Map<String, dynamic> body,
+      dynamic body,
+      Map<String, String>? headers,
+      CancelToken? cancelToken,
+      ResponseType responseType = ResponseType.map_data,
+      Function(int, int)? onSendProgress,
+      Function(int, int)? onReceiveProgress,
+      bool ignoreOnAllError = false,
       int urlIndex = 0}) async {
     String error = _checkBeforeRequest(isAuth, urlIndex);
     if (error.isNotEmpty) {
-      throw EticonApiError(error: error);
+      throw APIException(-1, body: error);
     }
-    return await ApiST.instance
-        .request(type: TYPE.PUT, method: method, isAuth: isAuth, testMode: testMode, query: body, urlIndex: urlIndex);
+    bool isUrl = path.startsWith('http');
+    return await DioApiST.instance.request(
+        type: 'PUT',
+        method: !isUrl ? path : null,
+        baseUrl: isUrl ? path : null,
+        isAuth: isAuth,
+        testMode: testMode,
+        headers: headers,
+        responseType: responseType ?? ResponseType.map_data,
+        cancelToken: cancelToken,
+        data: body,
+        onReceiveProgress: onReceiveProgress,
+        onSendProgress: onSendProgress,
+        ignoreOnAllError: ignoreOnAllError,
+        urlIndex: urlIndex);
   }
 
   /// Sends an HTTP DELETE request.
-  static Future<Map<String, dynamic>> delete(
-      {required String method,
-      bool isAuth = false,
+  static Future<dynamic> delete(String path,
+      {bool isAuth = false,
       bool testMode = false,
       Map<String, dynamic>? query,
+      Map<String, String>? headers,
+      CancelToken? cancelToken,
+      ResponseType responseType = ResponseType.map_data,
+      bool ignoreOnAllError = false,
       int urlIndex = 0}) async {
     String error = _checkBeforeRequest(isAuth, urlIndex);
     if (error.isNotEmpty) {
-      throw EticonApiError(error: error);
+      throw APIException(-1, body: error);
     }
-    return await ApiST.instance
-        .request(type: TYPE.DEL, method: method, isAuth: isAuth, testMode: testMode, query: query, urlIndex: urlIndex);
+    bool isUrl = path.startsWith('http');
+    return await DioApiST.instance.request(
+        type: 'DELETE',
+        method: !isUrl ? path : null,
+        baseUrl: isUrl ? path : null,
+        isAuth: isAuth,
+        testMode: testMode,
+        headers: headers,
+        responseType: responseType,
+        cancelToken: cancelToken,
+        data: query,
+        ignoreOnAllError: ignoreOnAllError,
+        urlIndex: urlIndex);
   }
 
   /// Sends an HTTP PATCH request.
-  static Future<Map<String, dynamic>> patch(
-      {required String method,
-      bool isAuth = false,
+  static Future<dynamic> patch(String path,
+      {bool isAuth = false,
       bool testMode = false,
-      required Map<String, dynamic> body,
+      dynamic body,
+      Map<String, String>? headers,
+      CancelToken? cancelToken,
+      ResponseType responseType = ResponseType.map_data,
+      Function(int, int)? onSendProgress,
+      Function(int, int)? onReceiveProgress,
+      bool ignoreOnAllError = false,
       int urlIndex = 0}) async {
     String error = _checkBeforeRequest(isAuth, urlIndex);
     if (error.isNotEmpty) {
-      throw EticonApiError(error: error);
+      throw APIException(-1, body: error);
     }
-    return await ApiST.instance
-        .request(type: TYPE.PATCH, method: method, isAuth: isAuth, testMode: testMode, query: body, urlIndex: urlIndex);
-  }
-
-  /// Sends an HTTP GET request.
-  static Future<Map<String, dynamic>> rawGet(
-      {required String url,
-      Map<String, String> headers = const {"Content-type": 'application/json'},
-      bool testMode = false,
-      Map<String, dynamic>? query,
-      int urlIndex = 0}) async {
-    return await ApiST.instance.request(
-        type: TYPE.GET, baseUrl: url, rawHeaders: headers, testMode: testMode, query: query, urlIndex: urlIndex);
-  }
-
-  /// Sends an HTTP POST request.
-  static Future<Map<String, dynamic>> rawPost(
-      {required String url,
-      Map<String, String> headers = const {"Content-type": 'application/json'},
-      bool testMode = false,
-      required Object body,
-      int urlIndex = 0}) async {
-    return await ApiST.instance.request(
-        type: TYPE.POST, baseUrl: url, rawHeaders: headers, testMode: testMode, query: body, urlIndex: urlIndex);
-  }
-
-  /// Sends an HTTP PUT request.
-  static Future<Map<String, dynamic>> rawPut(
-      {required String url,
-      Map<String, String> headers = const {"Content-type": 'application/json'},
-      bool testMode = false,
-      required Map<String, dynamic> body,
-      int urlIndex = 0}) async {
-    return await ApiST.instance.request(
-        type: TYPE.PUT, baseUrl: url, rawHeaders: headers, testMode: testMode, query: body, urlIndex: urlIndex);
-  }
-
-  /// Sends an HTTP DELETE request.
-  static Future<Map<String, dynamic>> rawDelete(
-      {required String url,
-      Map<String, String> headers = const {"Content-type": 'application/json'},
-      bool testMode = false,
-      Map<String, dynamic>? query,
-      int urlIndex = 0}) async {
-    return await ApiST.instance.request(
-        type: TYPE.DEL, baseUrl: url, rawHeaders: headers, testMode: testMode, query: query, urlIndex: urlIndex);
-  }
-
-  /// Sends an HTTP PUT request.
-  static Future<Map<String, dynamic>> rawPatch(
-      {required String url,
-      Map<String, String> headers = const {"Content-type": 'application/json'},
-      bool testMode = false,
-      required Map<String, dynamic> body,
-      int urlIndex = 0}) async {
-    return await ApiST.instance.request(
-        type: TYPE.PATCH, baseUrl: url, rawHeaders: headers, testMode: testMode, query: body, urlIndex: urlIndex);
+    bool isUrl = path.startsWith('http');
+    return await DioApiST.instance.request(
+        type: 'PATCH',
+        method: !isUrl ? path : null,
+        baseUrl: isUrl ? path : null,
+        isAuth: isAuth,
+        testMode: testMode,
+        headers: headers,
+        responseType: responseType,
+        cancelToken: cancelToken,
+        data: body,
+        onSendProgress: onSendProgress,
+        onReceiveProgress: onReceiveProgress,
+        ignoreOnAllError: ignoreOnAllError,
+        urlIndex: urlIndex);
   }
 
   static String _checkBeforeRequest(bool isAuth, int index) {
-    if (ApiST.instance.urls.isEmpty) {
+    if (DioApiST.instance.urls.isEmpty) {
       return 'Base url not set, use Api.init(String url)';
     }
-    if (ApiST.instance.urls.length <= index) {
+    if (DioApiST.instance.urls.length <= index) {
       return 'The index is greater than the number of urls in the list';
     }
     if (isAuth) {
@@ -292,3 +328,5 @@ class Api {
     return '';
   }
 }
+
+enum ResponseType { jsonResponse, stream, plain, byte, map_data }