+++ lib/src/old_api/dart_eticon_api.dart
@@ -1,5 +1,6 @@
+import 'package:eticon_api/src/old_api/type.dart';
+
 import 'api_st.dart';
-import 'type.dart';
 
 class DartApi {
   // ///Initialization API class
@@ -11,13 +12,13 @@ class DartApi {
   //       bool enableUtf8Decoding = false,
   //       String? storageUrl}) async {
   //   if (!ApiST.instance.setInitState()) {
-  //     throw EticonApiError(error: 'API class already initialization');
+  //     throw APIException(error: 'API class already initialization');
   //   }
   //   if (baseUrl.isEmpty) {
-  //     throw EticonApiError(error: 'URL is empty');
+  //     throw APIException(error: 'URL is empty');
   //   }
   //   if (!baseUrl.startsWith('http'))
-  //     throw EticonApiError(error: 'The url should start with https or http');
+  //     throw APIException(error: 'The url should start with https or http');
   //   if (baseUrl[baseUrl.length - 1] != '/') baseUrl += '/';
   //   ApiST.instance.setBaseUrl(baseUrl);
   //   ApiST.instance.setGlobalTestMode(globalTestMode);
@@ -31,7 +32,7 @@ class DartApi {
   // ///Help to get url to resource in server storage
   // static String dataFromStorage(String path) {
   //   if (ApiST.instance.storageUrl == null) {
-  //     throw EticonApiError(
+  //     throw APIException(
   //         error: 'Storage url is null, set storageUrl in Api.init()');
   //   } else {
   //     return '${ApiST.instance.storageUrl}$path';
@@ -50,11 +51,11 @@ class DartApi {
   //       bool testMode = false,
   //       Map<String, dynamic>? query}) async {
   //   if (ApiST.instance.baseUrl == null) {
-  //     throw EticonApiError(error: 'Base url not set, use Api.init()');
+  //     throw APIException(error: 'Base url not set, use Api.init()');
   //   }
   //   if (isAuth) {
   //     if (Token.instance.token.isEmpty) {
-  //       throw EticonApiError(
+  //       throw APIException(
   //           error:
   //           'Authentication token is empty, use Api.setToken (String url)');
   //     }
@@ -74,11 +75,11 @@ class DartApi {
   //       bool testMode = false,
   //       required Map<String, dynamic> body}) async {
   //   if (ApiST.instance.baseUrl == null) {
-  //     throw EticonApiError(error: 'Base url not set, use Api.init()');
+  //     throw APIException(error: 'Base url not set, use Api.init()');
   //   }
   //   if (isAuth) {
   //     if (Token.instance.token.isEmpty) {
-  //       throw EticonApiError(
+  //       throw APIException(
   //           error:
   //           'Authentication token is empty, use Api.setToken (String url)');
   //     }
@@ -98,11 +99,11 @@ class DartApi {
   //       bool testMode = false,
   //       required Map<String, dynamic> body}) async {
   //   if (ApiST.instance.baseUrl == null) {
-  //     throw EticonApiError(error: 'Base url not set, use Api.init()');
+  //     throw APIException(error: 'Base url not set, use Api.init()');
   //   }
   //   if (isAuth) {
   //     if (Token.instance.token.isEmpty) {
-  //       throw EticonApiError(
+  //       throw APIException(
   //           error:
   //           'Authentication token is empty, use Api.setToken (String url)');
   //     }
@@ -122,11 +123,11 @@ class DartApi {
   //       bool testMode = false,
   //       Map<String, dynamic>? query}) async {
   //   if (ApiST.instance.baseUrl == null) {
-  //     throw EticonApiError(error: 'Base url not set, use Api.init(String url)');
+  //     throw APIException(error: 'Base url not set, use Api.init(String url)');
   //   }
   //   if (isAuth) {
   //     if (Token.instance.token.isEmpty) {
-  //       throw EticonApiError(
+  //       throw APIException(
   //           error:
   //           'Authentication token is empty, use Api.setToken (String url)');
   //     }
@@ -143,10 +144,10 @@ class DartApi {
   /// Sends an HTTP GET request.
   static Future<Map<String, dynamic>> get(
       {required String url,
-        Map<String, String> headers = const {"Content-type": 'application/json'},
-        bool testMode = false,
-        Map<String, dynamic>? query,
-        int urlIndex = 0}) async {
+      Map<String, String> headers = const {"Content-type": 'application/json'},
+      bool testMode = false,
+      Map<String, dynamic>? query,
+      int urlIndex = 0}) async {
     return await ApiST.instance.request(
         type: TYPE.GET, baseUrl: url, rawHeaders: headers, testMode: testMode, query: query, urlIndex: urlIndex);
   }
@@ -154,10 +155,10 @@ class DartApi {
   /// Sends an HTTP POST request.
   static Future<Map<String, dynamic>> post(
       {required String url,
-        Map<String, String> headers = const {"Content-type": 'application/json'},
-        bool testMode = false,
-        required Object body,
-        int urlIndex = 0}) async {
+      Map<String, String> headers = const {"Content-type": 'application/json'},
+      bool testMode = false,
+      required Object body,
+      int urlIndex = 0}) async {
     return await ApiST.instance.request(
         type: TYPE.POST, baseUrl: url, rawHeaders: headers, testMode: testMode, query: body, urlIndex: urlIndex);
   }
@@ -165,10 +166,10 @@ class DartApi {
   /// Sends an HTTP PUT request.
   static Future<Map<String, dynamic>> put(
       {required String url,
-        Map<String, String> headers = const {"Content-type": 'application/json'},
-        bool testMode = false,
-        required Map<String, dynamic> body,
-        int urlIndex = 0}) async {
+      Map<String, String> headers = const {"Content-type": 'application/json'},
+      bool testMode = false,
+      required Map<String, dynamic> body,
+      int urlIndex = 0}) async {
     return await ApiST.instance.request(
         type: TYPE.PUT, baseUrl: url, rawHeaders: headers, testMode: testMode, query: body, urlIndex: urlIndex);
   }
@@ -176,10 +177,10 @@ class DartApi {
   /// Sends an HTTP DELETE request.
   static Future<Map<String, dynamic>> delete(
       {required String url,
-        Map<String, String> headers = const {"Content-type": 'application/json'},
-        bool testMode = false,
-        Map<String, dynamic>? query,
-        int urlIndex = 0}) async {
+      Map<String, String> headers = const {"Content-type": 'application/json'},
+      bool testMode = false,
+      Map<String, dynamic>? query,
+      int urlIndex = 0}) async {
     return await ApiST.instance.request(
         type: TYPE.DEL, baseUrl: url, rawHeaders: headers, testMode: testMode, query: query, urlIndex: urlIndex);
   }
@@ -187,10 +188,10 @@ class DartApi {
   /// Sends an HTTP PUT request.
   static Future<Map<String, dynamic>> patch(
       {required String url,
-        Map<String, String> headers = const {"Content-type": 'application/json'},
-        bool testMode = false,
-        required Map<String, dynamic> body,
-        int urlIndex = 0}) async {
+      Map<String, String> headers = const {"Content-type": 'application/json'},
+      bool testMode = false,
+      required Map<String, dynamic> body,
+      int urlIndex = 0}) async {
     return await ApiST.instance.request(
         type: TYPE.PATCH, baseUrl: url, rawHeaders: headers, testMode: testMode, query: body, urlIndex: urlIndex);
   }